Rename ChannelBase back to -> Channel, moving common methods to super class.
